backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6

# Machine Learning
scikit-learn==1.3.2
xgboost==2.0.2
shap==0.43.0
numpy==1.26.2
pandas==2.1.3
joblib==1.3.2
imbalanced-learn==0.11.0

# Utilities
python-dotenv==1.0.0

# Production
gunicorn==21.2.0
